+++ clang-tools-extra/include-cleaner/test/fragments-multi.cpp
@@ -1,0 +1,8 @@
+#include <vector>
+#include "a.inc"
+#include "b.inc"
+
+A AValue;
+B BValue;
+
+// RUN: clang-include-cleaner -print=changes %s --fragment-headers='.*\.inc$' -- -I%S/Inputs/ | count 0