+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -650,6 +650,12 @@ Miscellaneous
 Improvements to include-fixer
 -----------------------------
 
+Improvements to clang-include-cleaner
+-------------------------------------
+
+- Added :program:`clang-include-cleaner` support for treating matching direct
+  includes as fragments of the main file with ``--fragment-headers``.
+
 Improvements to clang-include-fixer
 -----------------------------------